labs/requirements.txt
psycopg2-binary==2.9.9
tabulate==0.9.0
